# cmd/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Noble blockchain indexer",
        "title": "Noble Indexer API",
        "termsOfService": "https://bakingbad.dev/terms",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/baking-bad/noble-indexer",
            "email": "hello@bakingbad.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/baking-bad/noble-indexer/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "noble.dipdup.net",
    "basePath": "/v1",
    "paths": {
...
        "/traces": {
            "get": {
                "description": "Returns a paginated list of execution traces showing internal calls, contract creations, and other EVM operations. Traces provide detailed insight into transaction execution. Can be filtered by transaction, addresses, contract, block height, or trace type.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List execution traces",
                "operationId": "list-transaction-traces",
                "parameters": [
                    {
                        "maxLength": 66,
                        "minLength": 66,
                        "type": "string",
                        "example": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef",
                        "description": "Filter by transaction hash (hexadecimal with 0x prefix)",
                        "name": "tx_hash",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of traces to return (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of traces to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by initiator address",
                        "name": "address_from",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by target address",
                        "name": "address_to",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by called contract address",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 12345,
                        "description": "Filter by block height",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "call",
                            "delegatecall",
                            "staticcall",
                            "create",
                            "create2",
                            "selfdestruct",
                            "reward",
                            "suicide"
                        ],
                        "type": "string",
                        "description": "Filter by trace type (comma-separated list)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order (default: desc)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of execution traces",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Trace"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
        "/tx": {
            "get": {
                "description": "Returns a paginated list of blockchain transactions. Can be filtered by addresses, contract, block height, transaction type, status, or time range. Supports various transaction types including legacy, EIP-1559 (dynamic fee), EIP-4844 (blob), and EIP-7702 (set code).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List transactions",
                "operationId": "list-transactions",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Number of transactions to return (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of transactions to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by sender address",
                        "name": "address_from",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by recipient address",
                        "name": "address_to",
                        "in": "query"
                    },
                    {
                        "maxLength": 42,
                        "minLength": 42,
                        "type": "string",
                        "example": "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb",
                        "description": "Filter by called contract address",
                        "name": "contract",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 12345,
                        "description": "Filter by block height",
                        "name": "height",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "TxTypeUnknown",
                            "TxTypeLegacy",
                            "TxTypeDynamicFee",
                            "TxTypeBlob",
                            "TxTypeSetCode"
                        ],
                        "type": "string",
                        "description": "Filter by transaction type (comma-separated list)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "TxStatusSuccess",
                            "TxStatusRevert"
                        ],
                        "type": "string",
                        "description": "Filter by execution status (comma-separated list)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1692892095,
                        "description": "Filter by timestamp from (Unix timestamp)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1692892095,
                        "description": "Filter by timestamp to (Unix timestamp)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order by timestamp (default: desc)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transactions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/tx/{hash}": {
            "get": {
                "description": "Returns detailed information about a specific transaction including status, gas used, value transferred, and associated traces",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction by hash",
                "operationId": "get-transaction",
                "parameters": [
                    {
                        "maxLength": 66,
                        "minLength": 66,
                        "type": "string",
                        "example": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef",
                        "description": "Transaction hash in hexadecimal with 0x prefix",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction information",
                        "schema": {
                            "$ref": "#/definitions/responses.Transaction"
                        }
                    },
                    "204": {
                        "description": "Transaction not found"
                    },
                    "400": {
                        "description": "Invalid transaction hash format",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "responses.Trace": {
            "description": "Transaction execution trace information",
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "1000000000000000000"
                },
                "contract": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000002"
                },
                "creation_method": {
                    "type": "string",
                    "example": "create"
                },
                "from_address": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000000"
                },
                "gas_limit": {
                    "type": "string",
                    "example": "21000"
                },
                "gas_used": {
                    "type": "string",
                    "example": "21000"
                },
                "height": {
                    "type": "integer",
                    "example": 100
                },
                "init_hash": {
                    "type": "string",
                    "example": "0x6060604052341561000f57600080fd5b"
                },
                "input": {
                    "type": "string",
                    "example": "0x0"
                },
                "output": {
                    "type": "string",
                    "example": "0x0"
                },
                "subtraces": {
                    "type": "integer",
                    "example": 0
                },
                "time": {
                    "type": "string",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "to_address": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000001"
                },
                "trace_address": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        0,
                        1
                    ]
                },
                "tx_hash": {
                    "type": "string",
                    "example": "0x0764012270afacd3b101bcfadaaa9fc3190d04ed90ff22"
                },
                "tx_position": {
                    "type": "integer",
                    "example": 0
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "call",
                        "delegatecall",
                        "staticcall",
                        "create",
                        "create2",
                        "selfdestruct"
                    ],
                    "example": "call"
                }
            }
        },
        "responses.Transaction": {
            "description": "Noble transaction information",
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 1000000000000000000
                },
                "cumulative_gas_used": {
                    "type": "integer",
                    "example": 21000
                },
                "effective_gas_price": {
                    "type": "integer",
                    "example": 1000000
                },
                "fee": {
                    "type": "string",
                    "example": "21000000000"
                },
                "from_address": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000000"
                },
                "gas": {
                    "type": "integer",
                    "example": 21000
                },
                "gas_price": {
                    "type": "integer",
                    "example": 1000000
                },
                "gas_used": {
                    "type": "integer",
                    "example": 21000
                },
                "hash": {
                    "type": "string",
                    "example": "0x0764012270afacd3b101bcfadaaa9fc3190d04ed90ff22c0ee59781e54858a7d"
                },
                "height": {
                    "type": "integer",
                    "example": 100
                },
                "index": {
                    "type": "integer",
                    "example": 0
                },
                "input": {
                    "type": "string",
                    "example": "0x"
                },
                "logs_bloom": {
                    "type": "string",
                    "example": "0x00000000000000000000000000000000000000000000"
                },
                "nonce": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "TxStatusSuccess",
                        "TxStatusRevert"
                    ],
                    "example": "TxStatusSuccess"
                },
                "time": {
                    "type": "string",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "to_address": {
                    "type": "string",
                    "example": "0x0000000000000000000000000000000000000001"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "TxTypeUnknown",
                        "TxTypeLegacy",
                        "TxTypeDynamicFee",
                        "TxTypeBlob",
                        "TxTypeSetCode"
                    ],
                    "example": "TxTypeDynamicFee"
                }
            }
        },
...
    }
}